Fix test suite compilation issues and build failures
- Fix Value.number() references using .int() instead across test files
- Resolve ambiguous 'Method' and 'Notification' references in MockTransport.swift
- Fix 'AnyRequest' type not found in MockTransport.swift
- Fix UIElement constructor in UIElementTests.swift (change description to elementDescription)
- Fix Value access methods in UIElementTests.swift
- Update AccessibilityTests.swift to fix AXUIElement type handling

Some tests still fail at runtime, but the code now compiles correctly.

+++ MacMCP/Tests/MacMCPTests/MenuInteractionTests.swift
@@ -49,7 +49,7 @@ struct MenuInteractionTests {
         // Use key commands to access menu (e.g., Alt+F for File menu)
         let input: [String: Value] = [
             "action": .string("press_key"),
-            "keyCode": .number(3)  // Example key code for F key
+            "keyCode": .int(3)  // Example key code for F key
         ]
         
         let result = try await tool.handler(input)
